requirements.txt
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
decorator==5.2.1
duckduckgo-search>=6.1.0
et_xmlfile==2.0.0
filelock==3.20.0
Flask==3.1.2
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
selenium==4.39.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.4
wsproto==1.3.2
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.3